requirements.txt
Flask==2.0.1
Flask_Cors==3.0.10
rpi-gpio==0.7.1
